Overload retrieve()
Set a normal function that returns a string.

The template function converts a value to another one. However, the implementation of the conversion
is error-prone when it is applied to a string. Indeed, castTo<String>() uses a string stream.
This stream ignores spaces from the original string.

In order to fix that, a specific function that just returns a string is defined. This function does not convert the string value.